+++ core/kover/learning/learners/__init__.py
@@ -16,5 +16,3 @@
 	You should have received a copy of the GNU General Public License
 	along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
-
-from ds import KoverDataset
