+++ core/setup.py
@@ -26,7 +26,7 @@ setup(
     url = "http://github.com/aldro61/kover",
     
     # Cython Extension
-    ext_modules = [Extension("kover/learning/set_covering_machine/popcount", ["kover/learning/set_covering_machine/popcount.c"], extra_compile_args=["-march=native"])],
+    ext_modules = [Extension("kover/learning/common/popcount", ["kover/learning/common/popcount.c"], extra_compile_args=["-march=native"])],
     data_files=[('kover/dataset/tools/kmer_tools/', ['kover/dataset/tools/kmer_tools/build/bin/dsk2kover', 
 																  'kover/dataset/tools/kmer_tools/build/bin/multidsk', 
 																  'kover/dataset/tools/kmer_tools/build/bin/dsk'])]
